+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v1.1.1 2025-03
 
+* [Add plugin function `openAttributePanel`](https://github.com/siyuan-note/siyuan/issues/14276)
+
 ## v1.1.0 2025-03-04
 
 * [Support floating window `Open in New Tab`](https://github.com/siyuan-note/siyuan/issues/11801)